cloudmgr: Move the package to pkg/adaptor/cloud
This patch changes the location of the cloudmgr package,
so that we can share this package with commands
other than cloud-api-adaptor.

Fixes #786
