+++ pkg/adaptor/cloud/cloudmgr/aws.go
