+++ pkg/adaptor/cloud/cloudmgr/azure.go
