+++ pkg/adaptor/cloud/cloudmgr/libvirt.go
