+++ pkg/adaptor/cloud/cloudmgr/manager.go
