+++ pkg/adaptor/cloud/cloudmgr/vsphere.go
